requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0

langchain==0.3.27
langchain-openai==0.2.8
langchain-text-splitters==0.3.11
langsmith==0.3.45
openai==1.54.4
tiktoken==0.8.0
httpx==0.27.2
requests==2.32.3
certifi==2025.10.5
distro==1.9.0

faiss-cpu==1.9.0
sentence-transformers==3.3.1
torch==2.5.1
numpy==1.26.4
scikit-learn==1.7.2
scipy==1.16.3

python-dotenv==1.0.1
pydantic==2.10.2
loguru==0.7.2
beautifulsoup4==4.12.3
lxml==5.3.0

streamlit==1.26.0
protobuf==4.25.8
pyarrow==22.0.0
pandas==2.3.3
altair==5.5.0
watchdog==6.0.0

diskcache>=5.6.3
datasets>=4.0.0
gitpython>=3.1.31
appdirs==1.4.4


streamlit==1.26.0
protobuf==4.25.8
pyarrow==22.0.0
pandas==2.3.3
altair==5.5.0
watchdog==6.0.0